+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Bob versions changelog
 
 
 #### Changed
+- Cleaner closes blobs instead of update ([#285](https://github.com/qoollo/bob/pull/300))
 - Only stable releases to latest tag on Docker Hub ([#339](https://github.com/qoollo/bob/pull/339))
 - Make Bob compile on stable toolchain.